# cm.enspd.educ-connect/openapi/main.yaml
---
openapi: "3.0.3"
info:
  title: "cm.enspd.educ-connect openApi"
  description: "Demo use open APi for generate REST API endpoints"
  version: "0.0.1"
servers:
- url: "https"
- url: "http"
paths:
  /demo:
    get:
      tags:
      - "Demo"
      summary: "Demo Summary"
      description: "Demo Description"
      operationId: "fetchAllDemo"
      security:
      - bearerAuth: []
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/DemoDTO"
        "401":
          description: "Unauthorized"
    post:
      tags:
      - "Demo"
      summary: "Create a new Demo"
      description: "Creates a new Demo with the provided data"
      operationId: "createDemo"
      security:
      - bearerAuth: []
      requestBody:
        description: "Demo object to be created"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemoDTO"
      responses:
        "201":
          description: "Demo created successfully"
          content:
            application/json:
              schema:
                type: "string"
                format: "uuid"
        "400":
          description: "Invalid input"
        "401":
          description: "Unauthorized"
  /register:
    post:
      tags:
      - "User"
      summary: "Register a new students"
      description: "Register a new students with the provided data"
      operationId: "registerStudent"
      security:
      - bearerAuth: []
      requestBody:
        description: "Demo object to be created"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserDTO"
      responses:
        "201":
          description: "Demo created successfully"
          content:
            application/json:
              schema:
                type: "string"
                format: "uuid"
        "400":
          description: "Invalid input"
        "401":
          description: "Unauthorized"
  /department:
    post:
      tags:
      - "Department"
      summary: "create new department"
      description: "create a new department"
      operationId: "createNewDepartment"
      security:
      - bearerAuth: []
      requestBody:
        description: "Level object to be created"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartment"
      responses:
        "201":
          description: "Level created successfully"
          content:
            application/json:
              schema:
                type: "string"
                format: "uuid"
                description: "identity of new department"
        "409":
          description: "department already exist"
        "401":
          description: "Unauthorized"
    get:
      tags:
      - "Department"
      summary: "get all department"
      description: "get all department"
      operationId: "getAllDepartment"
      security:
      - bearerAuth: []
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/DepartmentDTO"
        "401":
          description: "Unauthorized"
  /academic-training:
    post:
      tags:
      - "AcademicTraining"
      summary: "create new academic training"
      description: "create a new academic training"
      operationId: "createNewAcademicTraining"
      security:
      - bearerAuth: []
      requestBody:
        description: "Academic training object to be created"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAcademicTraining"
      responses:
        "201":
          description: "Academic training created successfully"
          content:
            application/json:
              schema:
                type: "string"
                format: "uuid"
                description: "identity of new academic training"
        "409":
          description: "academic training already exist"
        "401":
          description: "Unauthorized"
    get:
      tags:
      - "AcademicTraining"
      summary: "get all academic training"
      description: "get all academic training"
      operationId: "getAllAcademicTraining"
      security:
      - bearerAuth: []
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/AcademicTrainingDTO"
        "401":
          description: "Unauthorized"
  /level:
    post:
      tags:
      - "Level"
      summary: "create a new level"
      description: "create a new level with the provided data"
      operationId: "createNewLevel"
      security:
      - bearerAuth: []
      requestBody:
        description: "Level object to be created"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLevel"
      responses:
        "201":
          description: "Level created successfully"
          content:
            application/json:
              schema:
                type: "string"
                format: "uuid"
        "400":
          description: "Invalid input"
        "401":
          description: "Unauthorized"
    get:
      tags:
      - "Level"
      summary: "get all level"
      description: "get all level"
      operationId: "fetchLevels"
      security:
      - bearerAuth: []
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/LevelDTO"
        "401":
          description: "Unauthorized"
  /department/{departmentId}/program:
    post:
      tags:
      - "Department"
      summary: "add the new program"
      description: "add the new program to department"
      operationId: "addNewProgram"
      security:
      - bearerAuth: []
      parameters:
      - name: "departmentId"
        required: true
        in: "path"
        description: "the identity of the department"
        schema:
          type: "string"
          format: "uuid"
      requestBody:
        description: "Program object to be created"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProgram"
      responses:
        "201":
          description: "Program created successfully"
          content:
            application/json:
              schema:
                type: "string"
                format: "uuid"
        "409":
          description: "this program already exist in this department"
        "401":
          description: "Unauthorized"
    get:
      tags:
      - "Program"
      summary: "get program of the department"
      description: "get program of the department"
      operationId: "fetchAllProgramByDepartment"
      security:
      - bearerAuth: []
      parameters:
      - name: "departmentId"
        required: true
        in: "path"
        description: "the identity of the department"
        schema:
          type: "string"
          format: "uuid"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/ProgramDTO"
        "401":
          description: "Unauthorized"
        "404":
          description: "the department not found"
  /program/{programId}/speciality:
    post:
      tags:
      - "Program"
      summary: "add a new speciality"
      description: "add a new speciality with the provided data to program"
      operationId: "addNewSpeciality"
      security:
      - bearerAuth: []
      parameters:
      - name: "programId"
        in: "path"
        required: true
        description: "the identity of the program"
        schema:
          type: "string"
          format: "uuid"
      requestBody:
        description: "speciality object to be created"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpeciality"
      responses:
        "201":
          description: "Department created successfully"
          content:
            application/json:
              schema:
                type: "string"
                format: "uuid"
        "409":
          description: "the speciality already exist"
        "401":
          description: "Unauthorized"
    get:
      tags:
      - "Program"
      summary: "get all speciality of a program"
      description: "fetch all speciality of a program"
      operationId: "fetchAllSpecialityByProgram"
      security:
      - bearerAuth: []
      parameters:
      - name: "programId"
        in: "path"
        required: true
        description: "the identity of the program"
        schema:
          type: "string"
          format: "uuid"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/SpecialityDTO"
        "401":
          description: "Unauthorized"
components:
  schemas:
    DemoDTO:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        name:
          type: "string"
    UserDTO:
      type: "object"
      required:
      - "password"
      - "email"
      - "birthdate"
      - "level"
      - "contact"
      - "training"
      properties:
        password:
          type: "string"
        firstname:
          type: "string"
        lastname:
          type: "string"
        email:
          type: "string"
        username:
          type: "string"
        birthdate:
          type: "string"
          format: "date"
        contact:
          $ref: "#/components/schemas/ContactDTO"
        training:
          $ref: "#/components/schemas/AcademicTrainingDTO"
        level:
          type: "string"
          format: "uuid"
    ContactDTO:
      type: "object"
      properties:
        phoneNumbers:
          type: "array"
          items:
            $ref: "#/components/schemas/PhoneNumberDTO"
        whatsappNumber:
          $ref: "#/components/schemas/PhoneNumberDTO"
    PhoneNumberDTO:
      type: "object"
      properties:
        countryCode:
          type: "string"
          description: "the international code for the country"
          example: "237 for Cameroon"
        number:
          type: "string"
          description: "the phone number"
          example: 653492410
    AcademicTrainingDTO:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        name:
          type: "string"
    DepartmentDTO:
      type: "object"
      required:
      - "name"
      properties:
        id:
          type: "string"
          format: "uuid"
        name:
          type: "string"
        description:
          type: "string"
        program:
          $ref: "#/components/schemas/ProgramDTO"
    CreateSpeciality:
      type: "object"
      required:
      - "name"
      properties:
        name:
          type: "string"
        description:
          type: "string"
    SpecialityDTO:
      type: "object"
      required:
      - "name"
      properties:
        id:
          type: "string"
          format: "uuid"
        name:
          type: "string"
        description:
          type: "string"
        program:
          $ref: "#/components/schemas/ProgramDTO"
    ProgramDTO:
      type: "object"
      required:
      - "name"
      properties:
        id:
          type: "string"
          format: "uuid"
        name:
          type: "string"
        description:
          type: "string"
        department:
          $ref: "#/components/schemas/DepartmentDTO"
        level:
          $ref: "#/components/schemas/LevelDTO"
    CreateProgram:
      type: "object"
      required:
      - "name"
      properties:
        name:
          type: "string"
        description:
          type: "string"
    LevelDTO:
      type: "object"
      required:
      - "name"
      properties:
        id:
          type: "string"
          format: "uuid"
        name:
          type: "string"
        academicTraining:
          $ref: "#/components/schemas/AcademicTrainingDTO"
    CreateLevel:
      type: "object"
      required:
      - "name"
      properties:
        name:
          type: "string"
    CreateDepartment:
      type: "object"
      required:
      - "name"
      properties:
        name:
          type: "string"
          description: "name of department"
        description:
          type: "string"
    CreateAcademicTraining:
      type: "object"
      required:
      - "name"
      properties:
        name:
          type: "string"
        description:
          type: "string"